+++ crates/core/Cargo.toml
@@ -24,7 +24,6 @@ serde_json = { workspace = true }
 tokio = { workspace = true, features = ["macros", "rt", "sync", "time"] }
 tracing = { workspace = true }
 ureq = { workspace = true }
-sha2 = { workspace = true }
 
 [dev-dependencies]
 axum = { workspace = true, default-features = false, features = [
